validation-4

+++ dolphinscheduler-dao/src/main/java/org/apache/dolphinscheduler/dao/AlertDao.java
@@ -253,7 +253,7 @@ public class AlertDao {
                 .workflowInstanceId(workflowInstance.getId())
                 .workflowDefinitionCode(workflowInstance.getWorkflowDefinitionCode())
                 .workflowInstanceName(workflowInstance.getName())
-                .taskCode(taskInstance.getId())
+                .taskCode(taskInstance.getId().longValue())
                 .taskName(taskInstance.getName())
                 .taskType(taskInstance.getTaskType())
                 .taskStartTime(taskInstance.getStartTime())
